AI-service/requirements.txt
# Python AI Service Dependencies
# Requires CUDA-enabled GPU

# Core
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0

# PyTorch (install with CUDA support)
# For CUDA 11.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchvision>=0.15.0

# Stable Diffusion
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0
xformers>=0.0.22  # Memory efficient attention

# TripoSR dependencies
trimesh>=4.0.0
pygltflib>=1.16.0
einops>=0.7.0
omegaconf>=2.3.0
huggingface-hub>=0.19.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
rembg>=2.0.50  # Background removal
numpy>=1.24.0
scipy>=1.11.0

# 3D processing
open3d>=0.18.0
pymeshlab>=2023.12
